graph search form is on the market page, and the carosel displays stock data on graph
